+++ app/(root)/find-ride.tsx
@@ -1,5 +1,9 @@
+import CustomButton from "@/components/CustomButton";
+import GoogleTextInput from "@/components/GoogleTextInput";
 import RideLayout from "@/components/RideLayout";
+import { icons } from "@/constants";
 import { useLocationStore } from "@/store";
+import { router } from "expo-router";
 import React from "react";
 import { Text, View } from "react-native";
 
@@ -11,8 +15,34 @@ const FindRide = () => {
     setUserLocation,
   } = useLocationStore();
   return (
-    <RideLayout title="">
-      <Text className="text-2xl">Find Ride</Text>
+    <RideLayout title="Ride">
+      <View className="my-3">
+        <Text className="text-lg font-JakartaBold mb-3">From</Text>
+        <GoogleTextInput
+          icon={icons.target}
+          initialLocation={userAddress!}
+          containerStyle="bg-neutral-100"
+          textInputBackgroundColor="#f5f5f5"
+          handlePress={(location) => setUserLocation(location)}
+        />
+      </View>
+
+      <View className="my-3">
+        <Text className="text-lg font-JakartaBold mb-3">To</Text>
+        <GoogleTextInput
+          icon={icons.map}
+          initialLocation={destinationAddress!}
+          containerStyle="bg-neutral-100"
+          textInputBackgroundColor="transparent"
+          handlePress={(location) => setDestinationLocation(location)}
+        />
+      </View>
+
+      <CustomButton
+        title="Find now"
+        onPress={() => router.push("/(root)/confirm-ride")}
+        className="mt-5"
+      />
     </RideLayout>
   );
 };