+++ components/DriverCard.tsx
@@ -1,0 +1,10 @@
+import { Text, View } from "react-native";
+
+const DriverCard = () => {
+  return (
+    <View>
+      <Text>DriverCard</Text>
+    </View>
+  );
+};
+export default DriverCard;